[Audit][MODEL-01] Article/Post 코어 축소: 인바운드 AP 파싱·마크다운 추출
god node로 비대해진 Article/Post의 응집도 높은 관심사를 인터페이스 변경
없이 concern/functions로 이관.

Post (336→226줄):
- 인바운드 ActivityPub 파싱 class 메서드(from_activitypub_object 및
  reply/local host/body 추출 private 헬퍼) 전체를 Posts::FederationIngest
  concern으로 분리. 아웃바운드 직렬화 Posts::Federation과 대칭 구조.
- class_methods do 로 이관해 Post.from_activitypub_object 공개/그 외 private
  가시성 동일하게 유지 (호출부·테스트 변경 없음).

Article:
- I18n 위주의 to_markdown 표시 로직을 Articles::Markdown 함수 모듈로 이관,
  모델은 위임만 유지(.md 렌더 render markdown: 그대로 동작).

Refs #842 (MODEL-01)

테스트: bin/rails test (919 runs, 0 failures), bundle exec rspec (53, 0)
가시성 검증: Post.from_activitypub_object public / 나머지 private 유지 확인
RuboCop: 위반 없음
